chore: fix dir naming, drop unused config, fix compile errors
- rename 16_HashingTechs.cpp -> 16_HashingTechs and 13_Search Tree -> 13_SearchTree
- fix compile errors (missing includes, typos, unterminated comment)
- remove .nvmrc, .python-version, renovate.json (no JS/Python/deps here)
- note in README that numbering intentionally skips 04

+++ 10_Tree/01_BT_by_Array.cpp
@@ -1,5 +1,5 @@
 #include<iostream>
-#include<math>
+#include<cmath>
 using namespace std;
 
 class BT_Array{
@@ -7,7 +7,7 @@ class BT_Array{
         char *A;
     public:
         BT_Array(int a){
-            A = new char[pow(2,a)-1];
+            A = new char[(int)pow(2,a)-1];
             A = {0};
         }
         void insert(char a){
